pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
